create seperate controller and route file for changing password
